feat(client):use blueprint filter
